Fix issue where incorrect keys would be stored
Previously, there was an issue when using emplace_hint() where elements
would be added with an incorrect key when provided with a specific hint.
This has been fixed by carrying over the string iterator when calling
the common insert function NodeInTree().

Also added a unit test for testing the previously added insert()
overload.