Fix radix::extract for nodes with >1 children
radix::extract was broken when extracting nodes with more than 1
children. The reason was that a substitute node was created but never
linked to the parent, because the original node was blocking the space
inside the std::map. Now the original node emancipates itself from
their parent before adoption of the substitute.