Use radix::extract inside radix::erase
The only difference between extract() and erase() is that after erase()
is called, the node pointer is deleted. This can be realised by calling
extract() and letting the scope of the node handle run out. This will
cause deletion of the node pointer.

This fixes an issue with the previous erase() implementation. Whenever a
node that had any children was deleted, it would be made inaccessible
by changing it from a leaf to a non-leaf node. The object stored by this
node would remain in the tree but inaccessible to the user.

The previous function that handled the erasure was deleted, as it had
longer any use.

Additionally, the erase-unit test was rewritten to use assert calls.