Add const qualifier to Node::m_is_leaf
Made member const to enforce that nodes should never be
"made invisible" to the user. Instead, nodes should be exchanged for new
nodes, if a node changes from non-leaf to leaf.

This change required removing move constructor and assignment operator.
They were implemented but not in use, because Node is only every used as
pointer.